+++ cmd/agentchat/plugin_web.go
@@ -1,5 +1,0 @@
-//go:build !no_web
-
-package main
-
-import _ "github.com/Renaissance-Mind/agent-in-chat-feishu/web"